+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 ***********
 * [BREAKING] Adds AAA protection to PHT endpoints
+* [Breaking] Upgraded `OSD`` to version `4.0.0`
 * Updated ODA to v11.1.0
 * Adds auto-panel creation `POST` endpoint based on cycle description
 * Adds separate `PUT` endpoint to update panels