+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 ***********
 * [BREAKING] Adds AAA protection to PHT endpoints
 * [Breaking] Upgraded `OSD`` to version `4.0.0`
+* [Breaking] Updates ODA version to 11.3.0
 * Updated ODA to v11.1.0
 * Adds auto-panel creation `POST` endpoint based on cycle description
 * Adds separate `PUT` endpoint to update panels